possible fix - relatorio ativos

+++ FinanceTracker/Data/DepositoPrazo.cs
@@ -6,11 +6,11 @@ namespace FinanceTracker.Data
     {
         public int Id { get; set; }
         public int AtivoId { get; set; }
-        public float Valor { get; set; }
+        public double Valor { get; set; }
         public string Banco { get; set; }
         public string NumeroConta { get; set; }
         public string Titulares { get; set; }
-        public float TaxaJuroAnual { get; set; }
+        public double TaxaJuroAnual { get; set; }
         
         public AtivoFinanceiro AtivoFinanceiro { get; set; }
