+++ FinanceTracker/Data/ImovelArrendado.cs
@@ -8,10 +8,10 @@ namespace FinanceTracker.Data
         public int AtivoId { get; set; }
         public string Designacao { get; set; }
         public string Localizacao { get; set; }
-        public float ValorImovel { get; set; }
-        public float ValorRenda { get; set; }
-        public float ValorCondominio { get; set; }
-        public float OutrasDespesas { get; set; }
+        public double ValorImovel { get; set; }
+        public double ValorRenda { get; set; }
+        public double ValorCondominio { get; set; }
+        public double OutrasDespesas { get; set; }
 
         public AtivoFinanceiro AtivoFinanceiro { get; set; }
